sim: the path goal and the mission/scenario clocks, found by auditing outside sim.h
Asked a third time whether the type work was finished. It was not, and the
reason is that every audit so far had looked at sim.h. Three more, all in files
that hold sim state and none of which had ever been checked:

  * PathService's goalX/goalZ -> Fixed. These carry the EXACT destination
    through the search (the search itself works in cells and never reads them),
    and the caller round-tripped them float -> Fixed to build the final
    waypoint. That is the precision loss the conversion exists to remove, on
    the one value in a route that is not cell-aligned. The done() callback
    widens with them.
  * MissionScript::clock_ -> int32 ticks. It was `clock_ += dt` driving spawn
    timers and the victory/death timer conditions, and one caller already wrote
    `int32_t(clock_ * 30.0f)` to get ticks back out of it.
  * Scenario::clock_ -> int32 ticks, same shape, driving the Gametime
    conditions.

Both clocks are the last `+= dt` accumulators in the sim.

All 15 binaries pass, including missiontool, which is what actually exercises
the mission clock. Harness hash 8f42b8ce04e60a86, unchanged over two runs --
expected, since Inner Circle runs no mission script and the goal round trip was
lossless at these coordinates.

+++ src/sim/mission.h
@@ -121,7 +121,7 @@ public:
     void setIdents(std::unordered_map<std::string, int> m) { idents_ = std::move(m); }
 private:
 
-    float clock_ = 0;         // mission time (s), for timer conditions
+    int32_t clock_ = 0;       // mission time in TICKS, for timer conditions
     int outcome_ = 0;         // 0 / +1 / -1
     bool started_ = false;
 };
